+++ src/lib.rs
@@ -4,8 +4,10 @@
 #![doc = include_str!("../README.md")]
 
 pub mod orientation;
+pub mod position;
 
 /// The most commonly useful bits of the library
 pub mod prelude {
-    pub use {crate::orientation::Direction, crate::orientation::Rotation};
+    pub use crate::orientation::{Direction, Rotation};
+    pub use crate::position::Position;
 }