+++ RELEASES.md
@@ -2,4 +2,7 @@
 
 ## Version 0.1
 
-- Added `Direction` (a normalized `Vec2`) and `Rotation` (an angle from midnight) to make it easier to work with rotations in 2D
+### Enhancements
+
+- Added `Direction` (a normalized `Vec2`) and `Rotation` (an angle from midnight) to make it easier to work with rotations in 2D.
+- Added a `Position<T>` 2-dimensional coordinate type.